remove spare monthly goals feature
