--- src/extension/features/budget/display-monthly-goals-overview/index.css
@@ -1,23 +1,0 @@
-.tk-monthly-goals-overview-develop .colorize-currency .currency.positive {
-  color: var(--budget_balance_positive_background);
-}
-
-.tk-monthly-goals-overview-develop .colorize-currency .currency.negative {
-  color: var(--budget_balance_negative_background);
-}
-
-.tk-monthly-goals-overview-develop .total-income.warning .currency {
-  color: var(--budget_balance_warning_background);
-}
-
-.tk-monthly-goals-overview-develop .goal-remaining-balance.positive .currency {
-  color: var(--budget_balance_positive_background);
-}
-
-.tk-monthly-goals-overview-develop .goal-remaining-balance.negative .currency {
-  color: var(--budget_balance_negative_background);
-}
-
-.tk-monthly-goals-overview-develop .extra-bottom-space {
-  margin-bottom: 1rem;
-}
