+++ src/extension/features/budget/display-total-monthly-goals/FormattedCurrency.jsx
