+++ src/extension/features/budget/display-total-monthly-goals/InspectorCard.jsx
