+++ src/extension/features/budget/display-total-monthly-goals/index.css
@@ -1,0 +1,23 @@
+.tk-monthly-goals-overview .colorize-currency .currency.positive {
+  color: var(--budget_balance_positive_background);
+}
+
+.tk-monthly-goals-overview .colorize-currency .currency.negative {
+  color: var(--budget_balance_negative_background);
+}
+
+.tk-monthly-goals-overview .total-income.warning .currency {
+  color: var(--budget_balance_warning_background);
+}
+
+.tk-monthly-goals-overview .goal-remaining-balance.positive .currency {
+  color: var(--budget_balance_positive_background);
+}
+
+.tk-monthly-goals-overview .goal-remaining-balance.negative .currency {
+  color: var(--budget_balance_negative_background);
+}
+
+.tk-monthly-goals-overview .extra-bottom-space {
+  margin-bottom: 1rem;
+}
